+++ src/redux/digitalBills/reducer.js
@@ -1,15 +1,15 @@
-import { completeReducer, createReducer } from 'redux-recompose';
+// import { completeReducer, createReducer } from 'redux-recompose';
 
-import { actions } from './actions';
+// import { actions } from './actions';
 
-export const defaultState = {
-  digitalBillsLoading: null,
-  digitalBills: { mensaje: '' },
-  digitalBillsError: null
-};
+// export const defaultState = {
+//   digitalBillsLoading: null,
+//   digitalBills: { mensaje: '' },
+//   digitalBillsError: null
+// };
 
-const reducerDescription = {
-  primaryActions: [actions.PUT_DIGITAL_BILLS, actions.DELETE_DIGITAL_BILLS, actions.POST_DIGITAL_BILLS]
-};
+// const reducerDescription = {
+//   primaryActions: [actions.PUT_DIGITAL_BILLS, actions.DELETE_DIGITAL_BILLS, actions.POST_DIGITAL_BILLS]
+// };
 
-export const reducer = createReducer(defaultState, completeReducer(reducerDescription));
+// export const reducer = createReducer(defaultState, completeReducer(reducerDescription));
